refact(domain-page): services
