+++ app/dashboard/src/pages/services/dialog/index.tsx
@@ -21,9 +21,9 @@ import { InboundType } from 'types/Inbounds';
 import { FC, useEffect, useState } from 'react';
 import { FormProvider, useForm } from 'react-hook-form';
 import { useTranslation } from 'react-i18next';
-import { AddIcon, DeleteIcon, EditIcon } from 'components/Dialog/Icons';
-import { DialogModalHeader } from 'components/Dialog/ModalHeader';
-import { DialogModalFooter } from 'components/Dialog/ModalFooter';
+import { AddIcon, DeleteIcon, EditIcon } from 'components/dialog/Icons';
+import { DialogModalHeader } from 'components/dialog/ModalHeader';
+import { DialogModalFooter } from 'components/dialog/ModalFooter';
 import { getDefaultValues } from './DefaultValues';
 import { schema, FormType } from './FormSchema';
 import { InboundsField } from './InboundsField';
