+++ app/dashboard/src/pages/services/index.tsx
@@ -1,27 +1,28 @@
+
 import { FC, useEffect } from 'react';
-import { Box, VStack } from '@chakra-ui/react';
 import { ServicesFilters } from 'components/ServicesFilters';
 import { ServicesTable } from 'components/ServicesTable';
 import { pages } from 'constants/Pages';
-import { ServiceDialog } from 'components/ServiceDialog';
+import { ServiceDialog } from 'components/service-dialog';
 import { useServices } from 'contexts/ServicesContext';
 import { useDashboard } from 'contexts/DashboardContext';
 import { DeleteServiceModal } from 'components/DeleteServiceModal';
+import Page from 'components/page';
 
 
-export const ServicesPage: FC = () => {
+const ServicesPage: FC = () => {
   useEffect(() => {
     useDashboard.getState().activatePage(pages.findIndex((page) => page.path === '/services'));
     useServices.getState().refetchServices();
   }, []);
   return (
-    <VStack justifyContent="space-between" maxH="100vh" p="2rem" rowGap={4}>
-      <Box w="full" >
-        <ServiceDialog />
-        <DeleteServiceModal />
-        <ServicesFilters />
-        <ServicesTable />
-      </Box>
-    </VStack>
+    <Page>
+      <ServiceDialog />
+      <DeleteServiceModal />
+      <ServicesFilters />
+      <ServicesTable />
+    </Page>
   )
 }
+
+export default ServicesPage;
